+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Kelda Change Log
 Up Next
 -------------
 
+Release 0.12.0
+-------------
+
 - Allow containers to run in `privileged` mode by setting the `privileged` flag
 in the Container object. This is necessary for containers to access devices on
 the host machine, such as `/dev/fuse`.